refactor: move host_health_check() into host.py
